LayoutTests/inspector/timeline/timeline-event-RequestAnimationFrame-expected.txt

Tests 'RequestAnimationFrame' Timeline event records.


== Running test suite: TimelineEvent.RequestAnimationFrame
-- Running test case: TimelineEvent.RequestAnimationFrame.requestAnimationFrame
Starting Capture...
requestAnimationFrame fired
Stopping Capture...
PASS: Should be 1 AnimationFrameRequested record.
DETAILS: number
PASS: ScriptTimelineRecord details should be the requestAnimationFrame identifier.
